[Podcasts] make Episode.duration a BigInteger

+++ mygpo/podcasts/models.py
@@ -579,7 +579,7 @@ class Episode(UUIDModel, TitleModel, DescriptionModel, LinkModel,
     guid = models.CharField(max_length=200, null=True)
     content = models.TextField()
     released = models.DateTimeField(null=True, db_index=True)
-    duration = models.PositiveIntegerField(null=True)
+    duration = models.BigIntegerField(null=True)
     filesize = models.BigIntegerField(null=True)
     mimetypes = models.CharField(max_length=200)
     podcast = models.ForeignKey(Podcast, on_delete=models.PROTECT)
